requirements.txt
open3d==0.18.0
wandb
trimesh==4.0.10
pytorch_msssim
torchmetrics
tqdm
imageio==2.37.0
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
scipy==1.13.1
plyfile==1.1
roma
einops==0.8.0
numpy==1.26.4
PyQt5==5.15.11
matplotlib
evo==1.11.0
python-pycg
Einops
rerun-sdk
git+https://github.com/eriksandstroem/evaluate_3d_reconstruction_lib.git@9b3cc08be5440db9c375cc21e3bd65bb4a337db7
git+https://github.com/pengchongH/gaussian_rasterizer_smallerGSradii.git